vpk - code organization
